+++ addons/scene_manager/plugin.gd
@@ -5,6 +5,7 @@ var menu: Node
 
 func _enter_tree():
 	add_autoload_singleton("SceneManager", "res://addons/scene_manager/scene_manager.tscn")
+	add_autoload_singleton("Scenes", "res://addons/scene_manager/scenes.gd")
 	menu = preload("res://addons/scene_manager/menu.tscn").instance()
 	menu.name = "Scene Manager"
 
@@ -12,5 +13,6 @@ func _enter_tree():
 
 func _exit_tree():
 	remove_autoload_singleton("SceneManager")
+	remove_autoload_singleton("Scenes")
 	remove_control_from_docks(menu)
 	menu.free()