+++ addons/scene_manager/scenes.gd
@@ -1,0 +1,8 @@
+#
+# Please do not edit anything in this script
+#
+# Just use the editor to change everything you want
+#
+extends Node
+
+var scenes: Dictionary = {"_ignore_list":["res://addons"],"_sections":["Character","Menu"],"scene1":{"sections":["Menu"],"value":"res://demo/scene1.tscn"},"scene2":{"sections":[],"value":"res://demo/scene2.tscn"},"scene3":{"sections":["Menu","Character"],"value":"res://demo/scene3.tscn"}}